Error while "build and test"
I believe the problem is related with the additional import that I have added to the "core.py" file. I have moved the additional functions to other files to see if this is the problem.
