+++ src/magnet/io.py
@@ -1,6 +1,7 @@
-from importlib.resources import path
 import pandas as pd
 import streamlit as st
+
+from importlib.resources import path
 
 
 # Adapted from
@@ -115,3 +116,34 @@ def load_metadata(material, excitation):
     with path('magnet.data', f'{material}_{excitation}.h5') as h5file:
         data, metadata = h5_load(h5file)
     return metadata
+
+
+# Core losses for interpolation: cannot be in core.py due to errors when testing
+def core_loss_DI_sinusoidal(freq, flux, material):
+    excitation_read = 'datasheet'
+    df = load_dataframe_point(material, excitation_read, freq, flux)
+    if not df.empty:
+        core_loss = float(df['Power_Loss'])
+
+    else:
+        core_loss = 0.0  # Not interpolable
+    return core_loss
+
+
+def core_loss_SI_sinusoidal(freq, flux, material):
+    excitation_read = 'sinusoidal'
+    df = load_dataframe_point(material, excitation_read, freq, flux)
+    if not df.empty:
+        core_loss = float(df['Power_Loss'])
+
+    else:
+        core_loss = 0.0  # Not interpolable
+    return core_loss
+
+
+def loss_interpolated(waveform, algorithm, **kwargs):
+    assert waveform.lower() in ('sinusoidal', 'triangular', 'trapezoidal', 'arbitrary'), f'Unknown waveform {waveform}'
+    assert algorithm in ('DI', 'SI'), f'Unknown algorithm {algorithm}'
+
+    fn = globals()[f'core_loss_{algorithm}_{waveform.lower()}']
+    return fn(**kwargs)
