+++ src/magnet/plots.py
@@ -224,3 +224,30 @@ def core_loss_multiple(
         fig.update_yaxes(type='log')
 
     st.plotly_chart(fig, use_container_width=True)
+
+
+# Points for the representation of the plots
+def cycle_points_sinusoidal(point):
+    cycle_list = np.linspace(0, 1, point)
+    flux_list = np.sin(np.multiply(cycle_list, np.pi * 2))
+    volt_list = np.cos(np.multiply(cycle_list, np.pi * 2))
+    return [cycle_list, flux_list, volt_list]
+
+
+def cycle_points_trapezoidal(duty_p, duty_n, duty_0):
+    if duty_p > duty_n:
+        volt_p = (1 - (duty_p - duty_n)) / -(-1 - (duty_p - duty_n))
+        volt_0 = - (duty_p - duty_n) / -(-1 - (duty_p - duty_n))
+        volt_n = -1  # The negative voltage is maximum
+        b_p = 1  # Bpk is proportional to the voltage, which is is proportional to (1-dp+dN) times the dp
+        b_n = -(-1 - duty_p + duty_n) * duty_n / ((1 - duty_p + duty_n) * duty_p)  # Prop to (-1-dp+dN)*dn
+    else:
+        volt_p = 1  # The positive voltage is maximum
+        volt_0 = - (duty_p - duty_n) / (1 - (duty_p - duty_n))
+        volt_n = (-1 - (duty_p - duty_n)) / (1 - (duty_p - duty_n))
+        b_n = 1  # Proportional to (-1-dP+dN)*dN
+        b_p = -(1 - duty_p + duty_n) * duty_p / ((-1 - duty_p + duty_n) * duty_n)  # Prop to (1-dP+dN)*dP
+    cycle_list = [0, 0, duty_p, duty_p, duty_p + duty_0, duty_p + duty_0, 1 - duty_0, 1 - duty_0, 1]
+    flux_list = [-b_p, -b_p, b_p, b_p, b_n, b_n, -b_n, -b_n, -b_p]
+    volt_list = [volt_0, volt_p, volt_p, volt_0, volt_0, volt_n, volt_n, volt_0, volt_0]
+    return [cycle_list, flux_list, volt_list]
